✨ Implement internal Indexer (#67)
* ⬆️  add Configuration for Pragma registry

* ✨ Implement database handler for processing and storing indexed Pragma events

- Add `DatabaseHandler` struct to manage database operations for indexed events.
- Introduce `NewSpotEntry` and `NewMainnetSpotEntry` structs for database insertion.
- Implement event processing logic with retry mechanism for database inserts.
- Create global `INTERNAL_INDEXER_TRACKER` for monitoring indexer status and metrics.
- Add `start_pragma_indexer` function to initialize the Pragma data indexer and handle events.

* 🔧 Refactor configuration and indexing logic

- Remove `indexer_url` from `Config` struct as indexing is now handled internally.
- Introduce new modules for `indexing` and `monitoring` to manage indexing tasks and metrics.
- Implement `pragma_indexing_monitor` to handle the lifecycle of the Pragma data indexer with retry logic.
- Update monitoring metrics to track indexed events and latest indexed block.
- Clean up unused indexer-related code and improve overall structure for better maintainability.

* ✨ Add Apibara API Key to configuration files

- Introduce `APIBARA_API_KEY` in both `.env.example` and `README.md` to support new API integration.
- Update documentation to reflect the removal of the Indexer Service URL, as indexing is now integrated into the monitoring service.

* :wrench: Fix configuration issue

* ✨ Add protobuf-compiler installation to Dockerfile and CI

* Update src/main.rs


* Update src/main.rs


* Enhance event processing and reorg handling

* Refactor reorg event deletion logic in DatabaseHandler

- Enhance the `delete_invalidated_events` function to handle both spot and future entries during reorgs.
- Introduce a nuclear option for deleting all events, with appropriate logging for both scenarios.
- Update monitoring metrics to reflect the latest indexed block after deletions.

* :sparkles: Add APIBARA API key to Config and update indexing logic

* Refactor DatabaseHandler and InternalIndexerTracker initialization

---------

+++ src/processing/api.rs
@@ -67,7 +67,7 @@ pub async fn process_data_by_pair(
         }
     };
 
-    let time_since_last_update = match raw_time_since_last_update(result.timestamp) {
+    let time_since_last_update = match raw_time_since_last_update(result.timestamp as u64) {
         Ok(time) => time,
         Err(e) => {
             tracing::warn!(
